refactor dump configuration
