+++ homematicip/base/helpers.py
@@ -1,5 +1,10 @@
 from datetime import datetime
 import codecs
+import re
+import json
+import logging
+
+LOGGER = logging.getLogger(__name__)
 
 
 def get_functional_channel(channel_type, js):
@@ -49,3 +54,54 @@ def bytes2str(b):
             b.__class__.__name__
         )
     )
+
+def handle_config(json_state : str, anonymize : bool) -> str:
+    if "errorCode" in json_state:
+        LOGGER.error(
+            "Could not get the current configuration. Error: %s",
+            json_state["errorCode"],
+        )
+        return None
+    else:
+        c = json.dumps(json_state, indent=4, sort_keys=True)
+        if anonymize:
+            # generate dummy guids
+            c = anonymizeConfig(
+                c,
+                "[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
+                "00000000-0000-0000-0000-{0:0>12}",
+            )
+            # generate dummy SGTIN
+            c = anonymizeConfig(c, "3014F711[A-Z0-9]{16}", "3014F711{0:0>16}")
+            # remove refresh Token
+            c = anonymizeConfig(
+                c, '"refreshToken": ?"[^"]+"', '"refreshToken": null'
+            )
+            # location
+            c = anonymizeConfig(
+                c, '"city": ?"[^"]+"', '"city": "1010, Vienna, Austria"'
+            )
+            c = anonymizeConfig(
+                c, '"latitude": ?"[^"]+"', '"latitude": "48.208088"'
+            )
+            c = anonymizeConfig(
+                c, '"longitude": ?"[^"]+"', '"longitude": "16.358608"'
+            )
+
+        return c
+
+def anonymizeConfig(config, pattern, format):
+    m = re.findall(pattern, config, flags=re.IGNORECASE)
+    if m == None:
+        return config
+    map = {}
+    i = 0
+    for s in m:
+        if s in map.keys():
+            continue
+        map[s] = format.format(i)
+        i = i + 1
+
+    for k, v in map.items():
+        config = config.replace(k, v)
+    return config
